// frontend/src/api/generated/models/placeRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Trippy API
 * OpenAPI spec version: v1
 */

export interface PlaceRead {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  type: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  region: string;
  /** @nullable */
  neighborhood?: string | null;
  /** @minLength 1 */
  description: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  hours?: string | null;
  /** @nullable */
  duration_minutes?: number | null;
  /** @nullable */
  price_range?: string | null;
  /** @nullable */
  rating?: number | null;
  tags?: string[];
  /** @nullable */
  seasonal_notes?: string | null;
  booking_required?: boolean;
}